Add a null check on the adapted ValueMap
If resource is a SyntheticResource (which happens when the invoking component is included via a <cq:include path="<a non-existent path>" />), an attempt to adaptTo(ValueMap.class) will result in a null value.  In this case, it makes more sense to fail gracefully, returning an empty results object instead of causing the NPE to bubble up as an ELException.