Check for any serverN folders before adding root workflow instances folder
The root folder was being added to the workflow instances folder before all the folders had been checked for serverN folders.  This was causing serverN
folders to be compared with dated folders causing a number format exception in the folder comparator.  This will ignore dated folders in the root if
there is a serverN folder found.