Migrated JCRNodeChangeEventHandler to the ResourceChangeListener API.
Unlike the other event listeners, this one exposes its path filter as
an externally visible configuration, so it can't be completely
refactored without breaking clients. Instead, it will use the legacy
method if the path LDAP filter config property is present and the new
method if a list of paths is specified instead.