Resolves #1872 use "mvn oakpal:webster" to update nodetypes.cnd, privileges.xml, and checklists.
- in your maven settings.xml file, create two new profiles:
webster-aem64 and webster-aem65. In each, set a "webster.segmentStore"
property with the absolute path to the
crx-quickstart/repository/segmentstore directory for a local AEM 6.4
and 6.5 server intallation, respectively.
- in the content module, run "mvn oakpal:webster -Pwebster-aem65" to
update the nodetypes.cnd and privileges.xml files with any AEM 6.5 definitions
referenced by qualified name by files in the filevault archive root
(src/main/content/jcr_root). This should be performed regularly as
existing content is changed, and as new content is added to the content package.
- in the oakpal-checks, run "mvn oakpal:webster -Pwebster-aem64" and "mvn
oakpal:webster -Pwebster-aem65" to keep the forced roots in the content classifications
checklists up to date. This will likely only need to be performed when
features are added to the checklists, or when service packs are
released.