Major rewrite of the filter to support testing and logical separation
The tests were rewritten as well. Tests are now much more explicit and are closer to an idealized form of unit testing.

Also added a dependency on AssertJ. The fluent API makes writing assertions much easier, and has the (huge) benefit of making the assertions more clear, as to what they are asserting.