Update mocking libraries
To make the unittests work with JDK11, I had to update mockito and
powermock to versions supporting it. Therefor I upgraded to the most
recent ones...

Turned out to be not the best idea, because now a lot of test cases fail
because of some changes in there.

* Had to adjust quite a number of matchers (org.mockito.Matchers.* ->
org.mockito.ArgumentMatchers.*)
* any(someClass.class) does not match "null", you have to use any().
* anyInt does not match Long

* Mockito complains about unnecessary mockings, and it's best to remove
them.

2 main issue remain:

* powermock
* a bunch of "NoScrMetadata found" exceptions during build with maven
(works in the IDE though).