Adding a servlet to handle Robots.txt (#2243)
* Adding a servlet to handle Robots.txt. The servlet can operate in one of two modes, either pulling the content from a property (both relative and absolute paths are handled), or it can generate the entire file based on the specified osgi config attributes.

Note: the servlet must be addressed on the page's jcr:content node, e.g. /content/we-retail/us/en/_jcr_content.robots.txt. I'm unsure of why this is the case since it is not necessary for the sitemap servlet, but my best guess is something related to how AEM/Sling handles xml extensions on pages vs txt.