2271 bugfix for contexthub request (#2295)
It seems that Jetty does some internal validation on the PrintWriter, and you are allowed to invoke flushBuffer() only if there is some body at all. This situation was affecting the AemEnvironmentIndicatorFilter only when then response did not contain a "html" in the content-type and had an empty body (which is indeed a rare situation).